mcp_client/requirements.txt
livekit-agents>=1.2.0
livekit-plugins-openai>=0.6.0
mcp>=1.0.0
openai>=1.0.0
